backend/requirements.txt
fastapi==0.115.6

pymongo==4.10.1
mysql-connector-python-9.2.0

openai==1.58.1

requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
lxml==5.3.0
ultimate-sitemap-parser==0.5
# spacy==3.8.4

python-dotenv==1.0.1